+++ rag_system/kotaemon_install_guide/docker-compose.yml_example
@@ -1,0 +1,48 @@
+services:
+  kotaemon:
+    build:
+      context: ./kotaemon
+      target: full
+    entrypoint: ["/bin/sh", "-c", "tail -f /dev/null"]
+    environment:
+      GRADIO_SERVER_NAME: 0.0.0.0
+      GRADIO_SERVER_PORT: 7860
+    ports:
+      - '7860:7860'
+    volumes:
+      - './kotaemon/flowsettings.py:/app/flowsettings.py'
+      - './kotaemon/libs:/app/libs'
+      - './kotaemon/ktem_app_data:/app/ktem_app_data'
+    depends_on:
+      - ollama
+      - qdrant
+
+
+  ollama:
+    image : ollama/ollama
+    container_name: ollama
+    volumes:
+      - './ollama/:/root/.ollama'
+    ports:
+      - '11434:11434'
+    deploy:
+      resources:
+        reservations:
+          devices:
+            - driver: nvidia
+              count: all
+              capabilities: [gpu]
+  
+  qdrant:
+    image: qdrant/qdrant:latest
+    restart: always
+    container_name: qdrant
+    ports:
+      - 6333:6333
+      - 6334:6334
+    expose:
+      - 6333
+      - 6334
+      - 6335
+    volumes:
+      - ./qdrant_data:/qdrant/storage
